requirements.txt
mypy==0.982
wemake-python-styleguide==0.17.0
pydantic==1.10.2
python-telegram-bot==13.14
httpx
